Changed logger code and server handlers code to see if it helps.
