+++ internal/logger/logger.go
@@ -2,7 +2,10 @@ package logger
 
 import (
 	"fmt"
+	"net/http"
+	"time"
 
+	"github.com/go-chi/chi/v5/middleware"
 	"go.uber.org/zap"
 )
 
@@ -18,4 +21,30 @@ func InitLogger() *zap.Logger {
 	//	}
 	//}()
 	return logger
+}
+
+func Logger(lg *zap.Logger) func(next http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		fn := func(w http.ResponseWriter, r *http.Request) {
+			ww := middleware.NewWrapResponseWriter(w, r.ProtoMajor)
+
+			t1 := time.Now()
+			defer func() {
+				lg.Info("Request info",
+					zap.String("URI", r.RequestURI),
+					zap.String("method", r.Method),
+					zap.Duration("time", time.Since(t1)),
+				)
+			}()
+			defer func() {
+				lg.Info("Response info",
+					zap.Int("status", ww.Status()),
+					zap.Int("size", ww.BytesWritten()),
+				)
+			}()
+
+			next.ServeHTTP(ww, r)
+		}
+		return http.HandlerFunc(fn)
+	}
 }
