+++ internal/server/api/server.go
@@ -58,11 +58,12 @@ func (s *Server) routes() {
 	lg := logger.InitLogger()
 
 	s.router.Route("/", func(r chi.Router) {
+		r.Use(logger.Logger(lg))
 		r.Get("/", s.IndexHandler)
 		r.Get("/value/{mtype}/{mname}", s.GetHandler(lg))
 		r.Post("/value/", s.GetHandlerJSON(lg))
 		r.Post("/update/", s.UpdateHandlerJSON(lg))
-		r.Post("/update/{mtype}/{mname}/{mvalue}", s.UpdateHandler)
+		r.Post("/update/{mtype}/{mname}/{mvalue}", s.UpdateHandler(lg))
 	})
 }
